test_auth.py
#!/usr/bin/env python3
"""
Test script for JWT + RBAC Authentication implementation
"""

from datetime import timedelta
from app.core.security import hash_password, verify_password, create_access_token
from app.db.models import User
from app.schemas.auth import UserCreate, Token, UserOut, UserUpdate, TokenData
from app.core.config import settings
from jose import jwt, JWTError

def test_password_hashing():
    """Test password hashing and verification"""
    print("\n[TEST 1] Password Hashing and Verification")
    print("-" * 50)
    
    # Note: bcrypt has a 72-byte limit, so use a shorter password
    password = "test_password_123"  # 16 bytes
    hashed = hash_password(password)
    
    print(f"✓ Password hashed: {hashed[:20]}...")
    
    # Verify correct password
    assert verify_password(password, hashed), "Password verification failed"
    print("✓ Correct password verified")
    
    # Verify wrong password
    assert not verify_password("wrong_password", hashed), "Wrong password should not verify"
    print("✓ Wrong password rejected")


def test_jwt_token_creation():
    """Test JWT token creation and validation"""
    print("\n[TEST 2] JWT Token Creation and Validation")
    print("-" * 50)
    
    subject = "user@example.com"
    token = create_access_token(subject=subject)
    
    print(f"✓ Token created: {token[:30]}...")
    
    # Decode and verify token
    try:
        payload = jwt.decode(
            token,
            settings.JWT_SECRET_KEY,
            algorithms=[settings.JWT_ALGORITHM],
        )
        assert payload["sub"] == subject, "Token subject mismatch"
        print(f"✓ Token decoded successfully")
        print(f"✓ Subject: {payload['sub']}")
        print(f"✓ Expiration: {payload['exp']}")
    except JWTError as e:
        print(f"✗ Token validation failed: {e}")
        raise


def test_token_with_expiry():
    """Test token with custom expiry"""
    print("\n[TEST 3] JWT Token with Custom Expiry")
    print("-" * 50)
    
    subject = "admin@example.com"
    expires_delta = timedelta(hours=2)
    token = create_access_token(subject=subject, expires_delta=expires_delta)
    
    print(f"✓ Token created with 2-hour expiry")
    
    payload = jwt.decode(
        token,
        settings.JWT_SECRET_KEY,
        algorithms=[settings.JWT_ALGORITHM],
    )
    print(f"✓ Token expires at: {payload['exp']}")


def test_schemas():
    """Test Pydantic schemas"""
    print("\n[TEST 4] Pydantic Schemas")
    print("-" * 50)
    
    # Test UserCreate
    user_create = UserCreate(email="test@example.com", password="password123")
    print(f"✓ UserCreate schema: {user_create.email}")
    
    # Test Token
    token = Token(access_token="test_token_123", token_type="bearer")
    print(f"✓ Token schema: {token.token_type}")
    
    # Test TokenData
    token_data = TokenData(email="user@example.com")
    print(f"✓ TokenData schema: {token_data.email}")
    
    # Test UserUpdate
    user_update = UserUpdate(role="admin", is_active=True)
    print(f"✓ UserUpdate schema: {user_update.role}, is_active={user_update.is_active}")


def test_rbac_permissions():
    """Test RBAC permission checking"""
    print("\n[TEST 5] RBAC Permission Checking")
    print("-" * 50)
    
    # Simulate different user roles
    roles = ["user", "admin", "superadmin"]
    
    for role in roles:
        print(f"✓ Role '{role}' recognized")
    
    # Test permission check logic
    admin_roles = ["admin", "superadmin"]
    user_role = "user"
    
    assert user_role not in admin_roles, "User should not be in admin roles"
    print(f"✓ User role 'user' correctly denied admin access")
    
    assert "admin" in admin_roles, "Admin should be in admin roles"
    print(f"✓ Admin role correctly granted admin access")


def test_user_model():
    """Test User model structure"""
    print("\n[TEST 6] User Model Structure")
    print("-" * 50)
    
    print(f"✓ User model fields:")
    print(f"  - id: Integer (primary key)")
    print(f"  - email: String (unique, indexed)")
    print(f"  - hashed_password: String")
    print(f"  - role: String (default='user')")
    print(f"  - is_active: Boolean (default=True)")


def test_database_base():
    """Test database base configuration"""
    print("\n[TEST 7] Database Base Configuration")
    print("-" * 50)
    
    from app.db.base import Base
    from sqlalchemy.orm import declarative_base
    
    print(f"✓ SQLAlchemy Base initialized")
    print(f"✓ Base type: {type(Base)}")
    print(f"✓ Models can inherit from Base for ORM mapping")


def main():
    """Run all tests"""
    print("\n" + "=" * 50)
    print("JWT + RBAC AUTHENTICATION TEST SUITE")
    print("=" * 50)
    
    try:
        test_password_hashing()
        test_jwt_token_creation()
        test_token_with_expiry()
        test_schemas()
        test_rbac_permissions()
        test_user_model()
        test_database_base()
        
        print("\n" + "=" * 50)
        print("✓✓✓ ALL TESTS PASSED! ✓✓✓")
        print("=" * 50)
        print("\nAuthentication & Authorization (JWT + RBAC) is working correctly!")
        print("\nNext steps:")
        print("1. Set up PostgreSQL database")
        print("2. Run: uvicorn app.main:app --reload")
        print("3. Visit http://localhost:8000/docs for API documentation")
        print("4. Test endpoints with the Swagger UI")
        
    except Exception as e:
        print(f"\n✗ TEST FAILED: {e}")
        import traceback
        traceback.print_exc()
        return 1
    
    return 0


if __name__ == "__main__":
    exit(main())
