requirements.txt
# --- Core API ---
fastapi==0.110.0
uvicorn[standard]==0.27.1

# --- Configuration & Validation ---
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# --- Database ---
sqlalchemy>=2.0.27,<3.0
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Authentication & Security ---
python-jose==3.3.0
passlib[bcrypt]>=1.7.4

# --- Background Jobs & Caching ---
celery==5.3.6
redis==5.0.1

# --- Observability ---
prometheus-client==0.20.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0

# --- AI / NLP ---
torch==2.10.0
transformers==4.40.2
sentence-transformers==2.7.0
faiss-cpu==1.13.2
pytesseract==0.3.13
pillow==10.3.0

# --- Utilities ---
httpx==0.27.0
python-multipart==0.0.9
